Tried testing things to fix the issue
I tried adding a test.html file and adjusting the go file to find a fix to my issue. Some images show up, but not on the 404 page.

So I copied the 404.html code into test.html, and it appears it may be an issue with the HTML code. Though I'm still not sure why the external CSS isn't working